testeBancoDeDados feito e testes anteriores ajustados, para melhorar a organização
