+++ testeBancoDeDados/Scirpt.txt
@@ -1,0 +1,181 @@
+3.3 CRIAÇÃO DAS TABELAS DE ACORDO COM OS ARQUIVOS CSV:
+
+CREATE TABLE operadoras (
+    Registro_ANS VARCHAR(255) PRIMARY KEY,
+    CNPJ VARCHAR(20),
+    Razao_Social VARCHAR(255),
+    Nome_Fantasia VARCHAR(255),
+    Modalidade VARCHAR(100),
+    Logradouro VARCHAR(255),
+    Numero VARCHAR(20),
+    Complemento VARCHAR(100),
+    Bairro VARCHAR(100),
+    Cidade VARCHAR(100),
+    UF VARCHAR(2),
+    CEP VARCHAR(10),
+    DDD VARCHAR(5),
+    Telefone VARCHAR(20),
+    Fax VARCHAR(20),
+    Endereco_eletronico VARCHAR(255),
+    Representante VARCHAR(255),
+    Cargo_Representante VARCHAR(255),
+    Regiao_de_Comercializacao VARCHAR(255),
+    Data_Registro_ANS DATE
+);
+
+CREATE TABLE demonstracoes_contabeis (
+    DATA DATE,
+    REG_ANS VARCHAR(255),
+    CD_CONTA_CONTABIL VARCHAR(20),
+    DESCRICAO VARCHAR(255),
+    VL_SALDO_INICIAL VARCHAR(50),
+    VL_SALDO_FINAL VARCHAR(50),
+    PRIMARY KEY (DATA, REG_ANS, CD_CONTA_CONTABIL),
+    FOREIGN KEY (REG_ANS) REFERENCES operadoras(Registro_ANS)
+);
+
+obs.: criei colunas de valor em formato VARCHAR para que o SQL não reclamasse da "," ao invés do "." nos valores das demonstrações contábeis. 
+Depois alterarei a tabela para numeric.
+
+ALTER TABLE demonstracoes_contabeis
+ALTER COLUMN VL_SALDO_INICIAL TYPE NUMERIC USING REPLACE(VL_SALDO_INICIAL, ',', '.')::NUMERIC,
+ALTER COLUMN VL_SALDO_FINAL TYPE NUMERIC USING REPLACE(VL_SALDO_FINAL, ',', '.')::NUMERIC;
+
+
+3.4 IMPORTAÇÃO DAS TABELAS:
+1º - importei relatorio_cadop_canceladas também, pois, nas demonstrações de 2023 e 2024 tinham registros de seguradoras canceladas.
+2º - rodei o código: FunçãoRemoverZerosEsquerda para que pudesse relacionar as colunas das tabelas corretamente.
+3º - usei os arquivos editados no postgreSQL para improtar os dados das demonstrações contábeis.
+
+copy operadoras (Registro_ANS, CNPJ, Razao_Social, Nome_Fantasia, Modalidade, Logradouro, Numero, Complemento, Bairro, Cidade, UF, CEP, DDD, Telefone, Fax, Endereco_eletronico, Representante, Cargo_Representante, Regiao_de_Comercializacao, Data_Registro_ANS)
+FROM 'C:\Users\Michael\Desktop\Relatorio_cadop_canceladas.csv'
+DELIMITER ';'
+CSV HEADER;
+
+copy operadoras (Registro_ANS, CNPJ, Razao_Social, Nome_Fantasia, Modalidade, Logradouro, Numero, Complemento, Bairro, Cidade, UF, CEP, DDD, Telefone, Fax, Endereco_eletronico, Representante, Cargo_Representante, Regiao_de_Comercializacao, Data_Registro_ANS)
+FROM 'C:\Users\Michael\Desktop\Relatorio_cadop.csv'
+DELIMITER ';'
+CSV HEADER;
+
+copy demonstracoes_contabeis (DATA, REG_ANS, CD_CONTA_CONTABIL, DESCRICAO, VL_SALDO_INICIAL, VL_SALDO_FINAL)
+FROM 'C:\Users\Michael\Desktop\1T2024.csv'
+DELIMITER ';'
+CSV HEADER;
+
+copy demonstracoes_contabeis (DATA, REG_ANS, CD_CONTA_CONTABIL, DESCRICAO, VL_SALDO_INICIAL, VL_SALDO_FINAL)
+FROM 'C:\Users\Michael\Desktop\2T2024.csv'
+DELIMITER ';'
+CSV HEADER;
+
+copy demonstracoes_contabeis (DATA, REG_ANS, CD_CONTA_CONTABIL, DESCRICAO, VL_SALDO_INICIAL, VL_SALDO_FINAL)
+FROM 'C:\Users\Michael\Desktop\3T2024.csv'
+DELIMITER ';'
+CSV HEADER;
+
+copy demonstracoes_contabeis (DATA, REG_ANS, CD_CONTA_CONTABIL, DESCRICAO, VL_SALDO_INICIAL, VL_SALDO_FINAL)
+FROM 'C:\Users\Michael\Desktop\4T2024.csv'
+DELIMITER ';'
+CSV HEADER;
+
+copy demonstracoes_contabeis (DATA, REG_ANS, CD_CONTA_CONTABIL, DESCRICAO, VL_SALDO_INICIAL, VL_SALDO_FINAL)
+FROM 'C:\Users\Michael\Desktop\1T2023.csv'
+DELIMITER ';'
+CSV HEADER;
+
+copy demonstracoes_contabeis (DATA, REG_ANS, CD_CONTA_CONTABIL, DESCRICAO, VL_SALDO_INICIAL, VL_SALDO_FINAL)
+FROM 'C:\Users\Michael\Desktop\2T2023.csv'
+DELIMITER ';'
+CSV HEADER;
+
+copy demonstracoes_contabeis (DATA, REG_ANS, CD_CONTA_CONTABIL, DESCRICAO, VL_SALDO_INICIAL, VL_SALDO_FINAL)
+FROM 'C:\Users\Michael\Desktop\3T2023.csv'
+DELIMITER ';'
+CSV HEADER;
+
+copy demonstracoes_contabeis (DATA, REG_ANS, CD_CONTA_CONTABIL, DESCRICAO, VL_SALDO_INICIAL, VL_SALDO_FINAL)
+FROM 'C:\Users\Michael\Desktop\4T2023.csv'
+DELIMITER ';'
+CSV HEADER;
+
+
+3.5 DESENVOLVIMENTO DE QUERY ANALÍTICA PARA RESPONDER:
+• Quais as 10 operadoras com maiores despesas em "EVENTOS/ SINISTROS CONHECIDOS OU
+AVISADOS DE ASSISTÊNCIA A SAÚDE MEDICO HOSPITALAR" no último trimestre?
+
+SELECT
+    o.Registro_ANS,
+    o.Razao_Social,  -- Substitui o.Nome_Fantasia por o.Razao_Social
+    SUM(dc.VL_SALDO_FINAL) AS Despesas_Trimestre
+FROM
+    demonstracoes_contabeis dc
+JOIN
+    operadoras o ON dc.REG_ANS = o.Registro_ANS
+WHERE
+    REPLACE(dc.DESCRICAO, ' ', '') = 'EVENTOS/SINISTROSCONHECIDOSOUAVISADOSDEASSISTÊNCIAASAÚDEMEDICOHOSPITALAR'
+    AND dc.DATA >= (SELECT MAX(DATA) - INTERVAL '3 months' FROM demonstracoes_contabeis)
+GROUP BY
+    o.Registro_ANS, o.Razao_Social
+ORDER BY
+    Despesas_Trimestre DESC
+LIMIT 10;
+
+Obs.: repare que nesse caso tive que fazer a busca pela String sem espaçamentos, pois, não localizei em um primeiro momento "EVENTOS/ SINISTROS CONHECIDOS OU
+AVISADOS DE ASSISTÊNCIA A SAÚDE MEDICO HOSPITALAR", pois, pesquisei com espaçamento diferente ('EVENTOS/SINISTROS CONHECIDOS OU AVISADOS DE ASSISTÊNCIA A SAÚDE MEDICO HOSPITALAR').
+
+RETORNO:
+
+1  BRADESCO SAÚDE S.A. - R$ 54.458.189.118,43
+2  SUL AMERICA COMPANHIA DE SEGURO SAÚDE - R$ 36.823.120.491,38
+3  AMIL ASSISTÊNCIA MÉDICA INTERNACIONAL S.A. - R$ 36.221.526.783,22
+4  NOTRE DAME INTERMÉDICA SAÚDE S.A. - R$ 16.517.040.433,17
+5  HAPVIDA ASSISTENCIA MEDICA S.A. - R$ 13.658.120.881,70
+6  CAIXA DE ASSISTÊNCIA DOS FUNCIONÁRIOS DO BANCO DO BRASIL - R$ 13.002.216.049,19
+7  UNIMED NACIONAL - COOPERATIVA CENTRAL - R$ 12.194.602.690,55
+8  PREVENT SENIOR PRIVATE OPERADORA DE SAÚDE LTDA - R$ 10.354.968.079,51
+9  UNIMED BELO HORIZONTE COOPERATIVA DE TRABALHO MÉDICO - R$ 9.327.409.970,83
+10 UNIMED SEGUROS SAÚDE S/A - R$ 8.372.053.724,04
+
+
+• Quais as 10 operadoras com maiores despesas nessa categoria no último ano?
+
+SELECT
+    o.Registro_ANS,
+    o.Razao_Social,
+    SUM(dc.VL_SALDO_FINAL) AS Despesas_Ano
+FROM
+    demonstracoes_contabeis dc
+JOIN
+    operadoras o ON dc.REG_ANS = o.Registro_ANS
+WHERE
+    (dc.DESCRICAO ILIKE '%despesa%' OR dc.DESCRICAO ILIKE '%despesas%')
+    AND dc.DESCRICAO NOT ILIKE '%reembolso%'
+    AND dc.DESCRICAO NOT ILIKE '%recuperação%'
+    AND dc.DESCRICAO NOT ILIKE '%(-)%'
+    AND dc.DESCRICAO NOT ILIKE '%créditos%'
+    AND dc.DESCRICAO NOT ILIKE '%crédito%'
+    AND dc.DATA >= (SELECT MAX(DATA) - INTERVAL '1 year' FROM demonstracoes_contabeis)
+GROUP BY
+    o.Registro_ANS, o.Razao_Social
+ORDER BY
+    Despesas_Ano DESC
+LIMIT 10;
+
+RETORNO:
+
+1  BRADESCO SAÚDE S.A. - R$ 361.558.708.203,55
+2  SUL AMERICA COMPANHIA DE SEGURO SAÚDE - R$ 245.766.697.278,39
+3  AMIL ASSISTÊNCIA MÉDICA INTERNACIONAL S.A. - R$ 241.446.024.897,88
+4  HAPVIDA ASSISTENCIA MEDICA S.A. - R$ 149.117.950.770,49
+5  NOTRE DAME INTERMÉDICA SAÚDE S.A. - R$ 142.796.846.989,61
+6  UNIMED NACIONAL - COOPERATIVA CENTRAL - R$ 113.584.932.321,48
+7  CAIXA DE ASSISTÊNCIA DOS FUNCIONÁRIOS DO BANCO DO BRASIL - R$ 79.006.437.440,34
+8  UNIMED BELO HORIZONTE COOPERATIVA DE TRABALHO MÉDICO - R$ 72.580.486.846,53
+9  PREVENT SENIOR PRIVATE OPERADORA DE SAÚDE LTDA - R$ 70.594.130.404,50
+10 PORTO SEGURO - SEGURO SAÚDE S/A - R$ 48.943.288.273,15
+
+Informações para acessar o banco de dados, caso queira:
+Nome do host do servidor do banco de dados: localhost
+Nome do banco de dados: db_teste
+Nome do usuário: nome_do_usuario
+Senha do usuário: TESTE1234
+Número da porta: 5432
